Move HRTree iteration methods to separate module
