+++ src/hrtree.rs
@@ -46,11 +46,11 @@ const MAX_CAPACITY: usize = 2 * B - 1;
 type InsertionTuple<K, V> = Option<(K, V, u64, Box<Node<K, V>>)>;
 
 #[derive(Debug, Default)]
-struct Node<K, V> {
-    keys: ArrayVec<K, MAX_CAPACITY>,
-    values: ArrayVec<V, MAX_CAPACITY>,
+pub(crate) struct Node<K, V> {
+    pub(crate) keys: ArrayVec<K, MAX_CAPACITY>,
+    pub(crate) values: ArrayVec<V, MAX_CAPACITY>,
     hashes: ArrayVec<u64, MAX_CAPACITY>,
-    children: Option<ArrayVec<Box<Node<K, V>>, { MAX_CAPACITY + 1 }>>,
+    pub(crate) children: Option<ArrayVec<Box<Node<K, V>>, { MAX_CAPACITY + 1 }>>,
     tree_hash: u64,
     tree_size: usize,
 }
@@ -263,7 +263,7 @@ impl<K, V> Node<K, V> {
 }
 
 pub struct HRTree<K, V> {
-    root: Box<Node<K, V>>,
+    pub(crate) root: Box<Node<K, V>>,
 }
 
 impl<K, V> Default for HRTree<K, V> {
@@ -566,115 +566,6 @@ impl<K, V> PartialEq for HRTree<K, V> {
 }
 
 impl<K, V> Eq for HRTree<K, V> {}
-
-impl<K: Hash + Ord, V: Hash> FromIterator<(K, V)> for HRTree<K, V> {
-    fn from_iter<T>(iter: T) -> Self
-    where
-        T: IntoIterator<Item = (K, V)>,
-    {
-        let mut tree = HRTree::new();
-        let mut items: Vec<_> = iter.into_iter().collect();
-        items.sort_by(|a, b| a.0.cmp(&b.0));
-        for (k, v) in items {
-            tree.insert(k, v);
-        }
-        tree
-    }
-}
-
-enum IntoIterItem<K, V> {
-    Node(Box<Node<K, V>>),
-    Element(K, V),
-}
-
-pub struct IntoIter<K, V> {
-    stack: Vec<IntoIterItem<K, V>>,
-}
-
-impl<K, V> Iterator for IntoIter<K, V> {
-    type Item = (K, V);
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.stack.pop() {
-            Some(IntoIterItem::Node(mut node)) => {
-                if let Some(mut children) = node.children {
-                    self.stack.push(IntoIterItem::Node(children.pop().unwrap()));
-                    while !node.keys.is_empty() {
-                        let k = node.keys.pop().unwrap();
-                        let v = node.values.pop().unwrap();
-                        self.stack.push(IntoIterItem::Element(k, v));
-                        let c = children.pop().unwrap();
-                        self.stack.push(IntoIterItem::Node(c));
-                    }
-                } else {
-                    while !node.keys.is_empty() {
-                        let k = node.keys.pop().unwrap();
-                        let v = node.values.pop().unwrap();
-                        self.stack.push(IntoIterItem::Element(k, v));
-                    }
-                }
-                self.next()
-            }
-            Some(IntoIterItem::Element(k, v)) => Some((k, v)),
-            None => None,
-        }
-    }
-}
-
-impl<K, V> IntoIterator for HRTree<K, V> {
-    type Item = (K, V);
-    type IntoIter = IntoIter<K, V>;
-    fn into_iter(self) -> Self::IntoIter {
-        IntoIter {
-            stack: vec![IntoIterItem::Node(self.root)],
-        }
-    }
-}
-
-pub struct Iter<'a, K, V> {
-    stack: Vec<(&'a Node<K, V>, usize)>,
-}
-
-impl<'a, K, V> Iterator for Iter<'a, K, V> {
-    type Item = (&'a K, &'a V);
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some((node, children_passed)) = self.stack.pop() {
-            if children_passed < node.keys.len() {
-                self.stack.push((node, children_passed + 1));
-            }
-            if children_passed <= node.keys.len() {
-                if let Some(children) = node.children.as_ref() {
-                    self.stack.push((&children[children_passed], 0));
-                }
-            }
-            if 0 < children_passed && children_passed <= node.keys.len() {
-                Some((
-                    &node.keys[children_passed - 1],
-                    &node.values[children_passed - 1],
-                ))
-            } else {
-                self.next()
-            }
-        } else {
-            None
-        }
-    }
-}
-
-impl<'a, K, V> IntoIterator for &'a HRTree<K, V> {
-    type Item = (&'a K, &'a V);
-    type IntoIter = Iter<'a, K, V>;
-    fn into_iter(self) -> Self::IntoIter {
-        Iter {
-            stack: vec![(&self.root, 0)],
-        }
-    }
-}
-
-impl<K, V> HRTree<K, V> {
-    pub fn iter(&self) -> Iter<K, V> {
-        self.into_iter()
-    }
-}
 
 impl<K: std::fmt::Debug, V: std::fmt::Debug> std::fmt::Debug for HRTree<K, V> {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
@@ -1055,25 +946,5 @@ mod tests {
             expected_hash ^= super::hash(&key, &value);
             assert_eq!(tree1.hash(&..), expected_hash);
         }
-    }
-
-    #[test]
-    fn test_iter() {
-        let mut rng = rand::rngs::StdRng::seed_from_u64(42);
-        let mut key_values = Vec::new();
-
-        // add some
-        for _ in 0..1000 {
-            let key: u64 = rng.gen::<u64>();
-            let value: u64 = rng.gen();
-            key_values.push((key, value));
-        }
-        let tree = HRTree::from_iter(key_values.clone());
-        key_values.sort();
-        assert_eq!(
-            tree.iter().map(|(&k, &v)| (k, v)).collect::<Vec<_>>(),
-            key_values
-        );
-        assert_eq!(tree.into_iter().collect::<Vec<_>>(), key_values);
     }
 }
